requirements_all.txt
adal==1.2.7
aiocache==0.12.3
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.8
aiosignal==1.3.2
alembic==1.14.0
annotated-types==0.7.0
anthropic==0.45.1
anyio==4.8.0
appdirs==1.4.4
APScheduler==3.10.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
attrs==25.1.0
Authlib==1.3.2
av==12.3.0
azure-common==1.1.28
azure-core==1.32.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
bitarray==3.0.0
black==24.8.0
bleach==6.2.0
blinker==1.9.0
boto3==1.35.53
botocore==1.35.99
bs4==0.0.2
build==1.2.2.post1
cache==1.0.3
cachelib==0.9.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.2
click==8.1.8
colbert-ai==0.2.21
colorclass==2.2.2
coloredlogs==15.0.1
comm==0.2.2
compressed_rtf==1.0.6
cryptography==44.0.0
ctranslate2==4.5.0
dataclasses-json==0.6.7
datasets==3.2.0
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docstring_parser==0.16
docx2txt==0.8
duckduckgo_search==7.3.2
durationpy==0.9
easygui==0.98.3
ebcdic==1.1.1
ecdsa==0.19.0
einops==0.8.0
email_validator==2.2.0
emoji==2.14.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
Events==0.5
executing==2.2.0
extract-msg==0.52.0
fake-useragent==1.5.1
fastapi==0.111.0
fastapi-cli==0.0.7
faster-whisper==1.0.3
fastjsonschema==2.21.1
filelock==3.17.0
filetype==1.2.0
Flask==3.1.0
Flask-Caching==2.3.0
Flask-Cors==5.0.0
flatbuffers==25.1.24
fonttools==4.55.6
fpdf2==2.8.2
fqdn==1.5.1
frozenlist==1.5.0
fs==2.4.16
fsspec==2024.9.0
ftfy==6.2.3
gcp-storage-emulator==2024.8.3
git-python==1.0.3
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.6
google-api-core==2.24.0
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-genai==1.0.0
google-generativeai==0.7.2
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
greenlet==3.1.1
grpc-google-iam-v1==0.14.0
grpcio==1.67.1
grpcio-status==1.62.3
grpcio-tools==1.62.3
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.5.2
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.32.0
isodate==0.7.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kubernetes==32.0.0
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.31
langchain-text-splitters==0.3.5
langdetect==1.0.9
langfuse==2.44.0
langsmith==0.1.147
lark==1.1.9
ldap3==2.9.1
lxml==5.3.0
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.0
matplotlib-inline==0.1.7
mdurl==0.1.2
microsoft-bing-autosuggest==1.0.0
microsoft-bing-customimagesearch==1.0.0
microsoft-bing-customwebsearch==1.0.0
microsoft-bing-entitysearch==1.0.0
microsoft-bing-imagesearch==1.0.0
microsoft-bing-newssearch==1.0.0
microsoft-bing-spellcheck==1.0.0
microsoft-bing-videosearch==1.0.0
microsoft-bing-visualsearch==1.0.0
microsoft-bing-websearch==1.0.0
milvus-lite==2.4.11
mistune==3.1.1
mmh3==5.1.0
monotonic==1.6
moto==5.0.27
mpmath==1.3.0
msoffcrypto-tool==5.4.2
msrest==0.7.1
msrestazure==0.6.4.post1
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
olefile==0.47
oletools==0.60.2
onnxruntime==1.20.1
open-webui==0.5.7
openai==1.60.1
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opensearch-py==2.7.1
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.15
overrides==7.7.0
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pcodedmp==1.2.6
peewee==3.17.8
peewee-migrate==1.12.2
pexpect==4.9.0
pgvector==0.3.5
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
posthog==3.10.0
primp==0.11.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.0
protobuf==4.25.6
psutil==6.1.1
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
py-partiql-parser==0.6.1
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.7.1
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pymdown-extensions==10.11.2
pymilvus==2.5.0
pymongo==4.10.1
PyMySQL==1.1.1
pypandoc==1.13
pyparsing==3.2.1
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.3.4
pytest-docker==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-iso639==2025.1.27
python-jose==3.3.0
python-json-logger==3.2.1
python-magic==0.4.27
python-multipart==0.0.18
python-oxmsg==0.0.1
python-pptx==1.0.0
python-socketio==5.11.3
pytube==15.0.0
pytz==2024.2
pyxlsb==1.0.10
PyYAML==6.0.2
pyzmq==26.2.1
qdrant-client==1.12.2
rank-bm25==0.2.2
RapidFuzz==3.11.0
rapidocr-onnxruntime==1.3.24
red-black-tree-mod==1.20
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
responses==0.25.6
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.22.3
rsa==4.9
RTFDE==0.1.2
s3transfer==0.10.4
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
Send2Trash==1.8.3
sentence-transformers==3.3.1
sentencepiece==0.2.0
shapely==2.0.6
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.6
SQLAlchemy==2.0.32
stack-data==0.6.3
starlette==0.37.2
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.21.0
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.1
triton==3.1.0
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
ujson==5.10.0
unstructured==0.15.9
unstructured-client==0.28.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.30.6
uvloop==0.21.0
validators==0.34.0
vertexai==1.71.1
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.2.1
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.18.3
youtube-transcript-api==0.6.3
zipp==3.21.0
